+++ interpolate.h
@@ -20,9 +20,9 @@ extern std::vector<double>b,c,d;
 
 //---Functions
 double polynome(double a, double b, double c,double d, double x, double xj);
-double get_xinterval_id(std::vector<double>x_arr, double x, int &id);
-double interpolate(int n, std::vector<double> x,std::vector<double> a,
-                   std::vector<double> &b, std::vector<double> &c, std::vector<double> &d);	
+int get_xinterval_id(std::vector<double>x_arr, double x, int &id);
+int interpolate(int n, std::vector<double> x,std::vector<double> a,
+                std::vector<double> &b, std::vector<double> &c, std::vector<double> &d);	
 
 
 }
